refactor(codegen): [STABLE] eliminate _isEmbedded()/_isEmbeddedOrRetro() in favor of direct capability checks (#38)
- Add _ptrBytes() helper based on _cap('usize') for correct pointer sizes
  (fixes: AVR pointers are 2 bytes, not 4 as _isEmbedded() assumed)
- Replace printf format checks with _cap('bits') < 32 (16-bit int detection)
- Replace OS feature checks with _cap('os') === false
- Replace async feature checks with _cap('async') !== 'libuv'
- Replace allocator checks with _cap('allocator') !== 'heap'
- Replace embedded-only checks with _cap('async') === 'libuv' (inverse)
- Update 2 test expected.c files (allocator-none tests on desktop now use
  correct %d/%zu instead of embedded %ld/%u)
- All 20 phases pass (1953 tests)

+++ test/cases/phase11/allocator-none/value-type-stack/expected.c
@@ -5,6 +5,6 @@ typedef struct { int32_t x; int32_t y; } Node;
 int main(void) {
     TSC_INIT();
     Node n = { .x = 0, .y = 0 };
-    printf("%ld\n", (long)n.x);
+    printf("%d\n", n.x);
     return 0;
 }
